Avoid pylint error too-many-lines in server.py
