+++ logdetective/remote_log.py
@@ -1,23 +1,15 @@
-import io
 import logging
-from typing import Union
 from urllib.parse import urlparse
 
 import aiohttp
-
-from logdetective.server.compressors import TextCompressor
-
 
 LOG = logging.getLogger("logdetective")
 
 
 class RemoteLog:
     """
-    Handles retrieval and compression of remote log files.
+    Handles retrieval of remote log files.
     """
-
-    LOG_FILE_NAME = "log.txt"
-    COMPRESSOR = TextCompressor()
 
     def __init__(self, url: str, http_session: aiohttp.ClientSession):
         """
@@ -39,39 +31,6 @@ class RemoteLog:
     async def content(self) -> str:
         """Content of the url."""
         return await self.get_url_content()
-
-    @classmethod
-    def zip_text(cls, text: str) -> bytes:
-        """
-        Compress the given text.
-
-        Returns:
-            bytes: Compressed text
-        """
-        return cls.COMPRESSOR.zip({cls.LOG_FILE_NAME: text})
-
-    async def zip_content(self) -> bytes:
-        """
-        Compress the content of the remote log.
-
-        Returns:
-            bytes: Compressed log content
-        """
-        content_text = await self.content
-        return self.zip_text(content_text)
-
-    @classmethod
-    def unzip(cls, zip_data: Union[bytes, io.BytesIO]) -> str:
-        """
-        Uncompress the zipped content of the remote log.
-
-        Args:
-            zip_data: Compressed data as bytes or BytesIO
-
-        Returns:
-            str: The decompressed log content
-        """
-        return cls.COMPRESSOR.unzip(zip_data)[cls.LOG_FILE_NAME]
 
     def validate_url(self) -> bool:
         """Validate incoming URL to be at least somewhat sensible for log files
