+++ tests/test_server_remote_log.py
@@ -2,7 +2,8 @@ import pytest
 import aiohttp
 import aioresponses
 
-from logdetective.server.remote_log import RemoteLog
+from logdetective.remote_log import RemoteLog
+from logdetective.server.compressors import RemoteLogCompressor
 
 
 @pytest.mark.asyncio
@@ -16,10 +17,11 @@ async def test_server_remote_log():
         )
 
         remote_log = RemoteLog("https://example.com/logs/123", http_session)
+        compressor = RemoteLogCompressor(remote_log)
         content = await remote_log.content
         assert content
         assert mock_response in content
-        zip_data = await remote_log.zip_content()
-        url_text = remote_log.unzip(zip_data)
+        zip_data = await compressor.zip_content()
+        url_text = compressor.unzip(zip_data)
         assert url_text
         assert mock_response in url_text
